requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers
